Finalize before pull request
