+++ .idea/misc.xml
@@ -6,4 +6,7 @@
     <file type="web" url="file://$PROJECT_DIR$/StudyHabitsMicroService" />
   </component>
   <component name="ProjectRootManager" version="2" languageLevel="JDK_16" project-jdk-name="16" project-jdk-type="JavaSDK" />
+  <component name="SuppressKotlinCodeStyleNotification">
+    <option name="disableForAll" value="true" />
+  </component>
 </project>
